@@ -1,18 +1,26 @@
-const db = require("../db_connection");
-const fetch = require('node-fetch')
-const fs = require('fs');
-const path = require('path')
-const { spawn } = require('child_process');
-var ProgressBar = require('progress');
+import db from '../db_connection'
+import fetch from 'node-fetch'
+import fs from 'fs'
+import path from 'path'
+import {spawn} from 'child_process'
+import ProgressBar from 'progress'
+import commandExists from 'command-exists'
+import { dirname } from 'path';
+import { fileURLToPath } from 'url'
 
 
-async function updateDb(area, classType, count) {
-    const Query = `INSERT INTO MuniLEIMS.statisticalreport
-    (area_name, class, amount_streetlight)
-    VALUES ('${area}', '${classType}', '${count}')`;
+const _dirname = dirname(fileURLToPath(import.meta.url));
 
-    db.query(Query, function (err, result) {
-        // console.log(err, result)
+
+
+
+async function updateDb(city, area, consumption, classType, count) {
+    const query = `INSERT INTO MuniLEIMS.statisticalreport
+    (city, area, consumption, class, amount_streetlight)
+    VALUES ('${city}', '${area}', ${consumption}, '${classType}', '${count}')`;
+
+    db.query(query, function (err, result) {
+        console.log(err)
     })
 }
 
@@ -21,8 +29,6 @@ async function scanFile(fromPath) {
       const stat = await fs.promises.stat( fromPath );
       if( !stat.isFile() ) return [null,null]
 
-
-    // console.log( "Scanning '%s'", fromPath );
     let b64file = fs.readFileSync(fromPath, {encoding: 'base64'});
 
     let res1 = await fetch("https://detect.roboflow.com/muni-leims/3?api_key=7nC40FBnAQ5Ymx8wBDHq", {
@@ -42,14 +48,17 @@ async function scanFile(fromPath) {
     return [null,null]
 }
 
-async function sliceImage(scriptPath, file, outFolder) {
-    const python = spawn('python', [scriptPath, file, outFolder], {stdio: 'ignore'});
+async function sliceImage(scriptPath, file, outFolder) {    
+    let py3_exists = null
+    try {
+        py3_exists = await commandExists('python3')
+    } catch {}
+    const python = spawn(py3_exists ? 'python3' : 'python', [scriptPath, file, outFolder], {stdio: 'ignore'});
     // python.stderr.pipe(process.stderr);
     // python.stdout.pipe(process.stdout);
     await new Promise( (resolve, reject) => {
         python.on('exit', (code) => {
             if (code != 0) {
-                res.status(500).send(`python execute error`);
                 reject()
             } else {
                 resolve()
@@ -58,27 +67,9 @@ async function sliceImage(scriptPath, file, outFolder) {
     } )
 }
 
-async function uploadImage(req, res) {
-
-
-    let filename = req.file.originalname
-    let area = req.body.area
-
-
-    let file = path.resolve(__dirname, `./py-slice/${filename}`)
-    let outFolder = path.resolve(__dirname, `./py-slice/out`)
-    let scriptPath = path.resolve(__dirname, `./py-slice/main.py`)
-    fs.writeFileSync(file, req.file.buffer)
-    res.status(200).send('Image received and will scanned...')
-    await sliceImage(scriptPath, file, outFolder)
-    console.log('done slicing')
- 
-
-
-    const files = await fs.promises.readdir( outFolder);
-
-
-    var bar = new ProgressBar('  Scanning images [:bar] :percent :current/:total :etas', {
+async function scanImages(files, outFolder, city, area, consumption) {
+    
+    var bar = new ProgressBar('  Scanning images [:bar] :percent :current/:total', {
         complete: '=',
         incomplete: ' ',
         width: 20,
@@ -89,9 +80,36 @@ async function uploadImage(req, res) {
         // Get the full paths
         const fromPath = path.join(outFolder, file );
         let [count, classType] = await scanFile(fromPath)
-        if (count && classType) await updateDb(area, classType, count)
+        console.log(city, area, consumption, classType, count)
+        if (count && classType) await updateDb(city, area, consumption, classType, count)
         bar.tick()
     }
 }
 
-module.exports = {uploadImage};
+export async function uploadImage(req, res) {
+
+
+    let filename = req.file.originalname
+    let area = req.body.area
+    let city = req.body.city
+    let consumption = req.body.consumption
+
+
+    let file = path.resolve(_dirname, `./py-slice/${filename}`)
+    console.log(`trying to open ${file}`)
+    let outFolder = path.resolve(_dirname, `./py-slice/out`)
+    let scriptPath = path.resolve(_dirname, `./py-slice/main.py`)
+    fs.writeFileSync(file, req.file.buffer)
+    try {
+        await sliceImage(scriptPath, file, outFolder)
+    } catch {
+        res.status(500).send(`python execute error`);
+    }
+    res.status(200).send('Image received and will scanned...')
+    console.log('done slicing')
+ 
+    const files = await fs.promises.readdir(outFolder);
+    await scanImages(files, outFolder, city, area, consumption)
+
+}
+
