@@ -1,20 +1,25 @@
 import React from 'react';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
-// import LoginPage from './routes/loginPage';
+import LoginPage from './routes/loginPage';
 import HomePage from './routes/homePage';
 import StatisticPage from './routes/statisticPage';
-// import { PrivateRoute } from './component/routing';
+// import { PrivateRoute } from './components/routing';
 import NotFoundPage from './routes/page404';
 import MapPage from './routes/mapPage';
 import ImgUpload from './routes/imageUpload'
-// import ConvertCurency from './component/convertCurency'
+import Reports from './routes/reports';
+
+// import ConvertCurency from './components/convertCurency'
 function App() {
   return (
     <Router>
       <Routes>
         <Route exact path="/homepage" element={<HomePage />} />
+        <Route exact path="/" element={<Reports dataType="cities" />}/>
         {/* <Route exact path="/" element={<LoginPage />} /> */}
-        <Route exact path="/" element={<ImgUpload />} />
+        <Route exact path="/ImgUpload" element={<ImgUpload />} />
+        <Route exact path="/municipalities" element={<Reports dataType='municipalities' />} />
+        <Route exact path="/cities" element={<Reports dataType='cities' />} />
         <Route exact path="/statisticPage" element={<StatisticPage />} />
         <Route exact path="/mappage" element={<MapPage />} />
         {/* <Route
