@@ -1,8 +1,10 @@
 import React from 'react';
 import sidebarConfig from './sidebarConfig';
-import { Header, Button, Menu } from '../component/util/board';
+import { Header, Button, Menu } from '../components/util/board';
+import { useNavigate } from 'react-router-dom';
 
 function HomePage() {
+  const navigate = useNavigate()
   return (
     <>
       <Header>
@@ -14,7 +16,11 @@ function HomePage() {
         <h2>Menu:</h2>
         <section>
           {sidebarConfig.map((navigation) => (
-            <Button key={navigation.title} sx={{ width: '100%' }}>
+            <Button 
+                onClick={() => navigate(navigation.path)} 
+                key={navigation.title} 
+                sx={{ width: '100%' }}
+            >
               <h2>{navigation.title} </h2>
             </Button>
           ))}
