@@ -1,8 +1,8 @@
 import React from 'react';
 import AccountCircleIcon from '@mui/icons-material/AccountCircle';
 import { TextField, Button } from '@mui/material';
-import Logo from '../component/images/logo.svg';
-import { Header, Login } from '../component/util/board';
+import Logo from '../components/images/logo.svg';
+import { Header, Login } from '../components/util/board';
 
 function LoginPage() {
   return (
