@@ -1,6 +1,10 @@
-module.exports = {
-  DB_HOST: process.env.DB_HOST,
-  DB_USER: process.env.DB_USER,
-  DB_PASS: process.env.DB_PASS,
-  DB_PORT: process.env.DB_PORT,
-};
+import dotenv from 'dotenv'
+dotenv.config()
+
+const DB_HOST = process.env.DB_HOST
+const DB_USER = process.env.DB_USER
+const DB_PASS = process.env.DB_PASS
+const DB_PORT = process.env.DB_PORT
+
+
+export { DB_HOST, DB_USER, DB_PASS, DB_PORT }
