@@ -1,9 +1,7 @@
-// eslint-disable-next-line import/no-extraneous-dependencies
-require('dotenv').config(); // comment this line before uploading to heroku
-require('./server');
-// eslint-disable-next-line import/extensions
-require('./db_connection.js');
-
+import dotenv from 'dotenv'
+dotenv.config()
+import './server'
+import './db_connection'
 
 
 
